backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.5
pydantic==2.9.2
SQLAlchemy==2.0.36
python-dateutil==2.9.0.post0
orjson==3.10.7
phonenumberslite==8.13.47
email-validator==2.2.0
